added object count at deleting; added confim window

+++ src/main/java/application/graphic/ui/DeleteEntity.java
@@ -14,19 +14,21 @@ public class DeleteEntity<T, E extends IEntityFrame<T>> extends ToolButton  {
     super("Удалить", "images/remove.png");
     this.addActionListener(new ActionListener() {
       public void actionPerformed(ActionEvent e) {
-        String message = "Вы действительно хотите удалить выбранную(ые) запись(и)?";
-        int result = JOptionPane.showConfirmDialog(null,
-                message,
-                "Подтверждение действия",
-                JOptionPane.YES_NO_OPTION,
-                JOptionPane.QUESTION_MESSAGE);
-        if (result == JOptionPane.YES_OPTION) {
-          EntityDao<T> dao = new EntityDao<>(entityClass);
-          List<T> objects = parentWindow.getObjectsToDelete();
-          for(T object: objects) {
-            dao.deleteObject(object);
+        List<T> objects = parentWindow.getObjectsToDelete();
+        if(objects.size() > 0) {
+          String message = "Вы действительно хотите удалить выбранную(ые) запись(и)?";
+          int result = JOptionPane.showConfirmDialog(null,
+                  message,
+                  "Подтверждение действия",
+                  JOptionPane.YES_NO_OPTION,
+                  JOptionPane.QUESTION_MESSAGE);
+          if (result == JOptionPane.YES_OPTION) {
+            EntityDao<T> dao = new EntityDao<>(entityClass);
+            for(T object: objects) {
+              dao.deleteObject(object);
+            }
+            parentWindow.setTable();
           }
-          parentWindow.setTable();
         }
       }
     });
